Reusable CoptBtn and adds CopyBtn to cards
