+++ src/components/CardComponent.jsx
@@ -1,5 +1,6 @@
 import React, { useState } from 'react';
 import CodePreview from './CodePreview';
+import CopyBtn from './CopyBtn';
 
 function CardComponent({ data }) {
 	const { prompt, component, score, name, upvotes, username, category } = data;
@@ -36,7 +37,8 @@ function CardComponent({ data }) {
 		<div className='flex flex-col justify-center items-center h-[420px]'>
 			<div className='relative flex flex-col items-center rounded-[20px] w-[400px] mx-auto p-4 bg-white bg-clip-border shadow-lg shadow-shadow-500'>
 				<div className='relative flex h-32 w-full justify-center rounded-xl bg-cover'>
-					<div className='absolute flex h-32 w-full justify-center rounded-xl bg-cover bg-gradient-to-tl from-green-50 to-green-300'>
+					<div className='absolute flex h-full w-full justify-center rounded-xl bg-cover bg-gradient-to-tl from-green-50 to-green-300'>
+						<CopyBtn copy={component} />
 						<CodePreview generatedCode={component} />
 					</div>
 					<div className='absolute -bottom-12 flex h-[87px] w-[87px] items-center justify-center rounded-full border-[4px] border-white bg-pink-400 dark:!border-navy-700'>
